Move config_parser.* into its own subfolder
